# Conflicts:
#	resources/views/admin/configuration.blade.php
#	resources/views/admin/modules/aiform/settings.blade.php

+++ resources/views/livewire/comments/comment-item.blade.php
@@ -8,7 +8,7 @@
                 {{ $editCommentId }} {{ Auth::check() }}
                 @if(!$editCommentId && Auth::check())
                     <div class="btn-group col-2">
-                        @if(Auth::user()->id === $comment->creator_id)
+                        @if((int)Auth::user()->id === (int)$comment->creator_id)
                             <button type="button" class="btn btn-primary btn-sm"
                                     wire:click.prevent="startEditingComment( {{ $comment->id }}, {{$comment}} )">
                                 {{__('Edit')}}
@@ -35,7 +35,7 @@
                 </div>
             </div>
 
-            @if($editCommentId === $comment->id)
+            @if((int)$editCommentId === (int)$comment->id)
 
                 <div class="flex-1 mb-3">
                     <textarea
@@ -53,7 +53,7 @@
                 </div>
             @endif
 
-            @if($replyCommentId === $comment->id)
+            @if((int)$replyCommentId === (int)$comment->id)
                 <div class="flex-1 mb-3 ml-3 mt-3">
                   <textarea
                       class="form-control mb-3"
